refactor!: promote Article and Enclosure out of RssBuilder
