+++ lib/html2rss/rendering/description_builder.rb
@@ -36,7 +36,7 @@ module Html2rss
       # @param base [String] The base text content for the description
       # @param title [String] The article title (used for alt text and title removal)
       # @param url [String, Html2rss::Url] The article URL (used for sanitization)
-      # @param enclosures [Array<Html2rss::RssBuilder::Enclosure>, nil] Media enclosures
+      # @param enclosures [Array<Html2rss::Article::Enclosure>, nil] Media enclosures
       # @param image [String, Html2rss::Url, nil] Fallback image URL
       def initialize(base:, title:, url:, enclosures:, image:)
         @base = base.to_s
