+++ spec/lib/html2rss/rss_builder_spec.rb
@@ -21,20 +21,20 @@ RSpec.describe Html2rss::RssBuilder do
   end
   let(:articles) do
     [
-      Html2rss::RssBuilder::Article.new(url: 'http://example.com/1',
-                                        id: 1,
-                                        title: 'Title 1',
-                                        description: 'Description 1',
-                                        published_at: '1969-12-31 23:59:59',
-                                        image: 'http://example.com/image1.jpg',
-                                        scraper: RSpec),
-      Html2rss::RssBuilder::Article.new(url: 'http://example.com/2',
-                                        id: 2,
-                                        title: 'Title 2',
-                                        description: 'Description 2',
-                                        published_at: '1969-12-31 23:59:59',
-                                        image: 'http://example.com/image2.jpg',
-                                        scraper: Html2rss::AutoSource::Scraper::Html)
+      Html2rss::Article.new(url: 'http://example.com/1',
+                            id: 1,
+                            title: 'Title 1',
+                            description: 'Description 1',
+                            published_at: '1969-12-31 23:59:59',
+                            image: 'http://example.com/image1.jpg',
+                            scraper: RSpec),
+      Html2rss::Article.new(url: 'http://example.com/2',
+                            id: 2,
+                            title: 'Title 2',
+                            description: 'Description 2',
+                            published_at: '1969-12-31 23:59:59',
+                            image: 'http://example.com/image2.jpg',
+                            scraper: Html2rss::AutoSource::Scraper::Html)
     ]
   end
